fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@21824 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/gui/tracker/TrackerValueDesc.cpp
@@ -48,7 +48,7 @@ TrackerValueDesc::TrackerValueDesc(const std::string& name,
                                    SUMOTime recordBegin)
     : myName(name), myActiveCol(col), myInactiveCol(col),
       myMin(0), myMax(0),
-      myAggregationInterval(TIME2STEPS(1) / DELTA_T), myInvalidValue(-1), myValidNo(0),
+      myAggregationInterval((int)(TIME2STEPS(1) / DELTA_T)), myInvalidValue(-1), myValidNo(0),
       myRecordingBegin(recordBegin), myTmpLastAggValue(0) {}
 
 
@@ -147,7 +147,7 @@ void
 TrackerValueDesc::setAggregationSpan(SUMOTime as) {
     AbstractMutex::ScopedLocker locker(myLock);
     if (myAggregationInterval != as / DELTA_T) {
-        myAggregationInterval = as / DELTA_T;
+        myAggregationInterval = (int)(as / DELTA_T);
         // ok, the aggregation has changed,
         //  let's recompute the list of aggregated values
         myAggregatedValues.clear();
